projets/api_rest_106977_2fbe73/requirements.txt
fastapi==0.92.2
uvicorn==0.17.6
motor==2.3.2
pymongo==3.12.4
python-jose==3.6.0
cryptography==3.4.8
pydantic==1.9.1
swagger-ui==4.11.1
requirements==2.5.0.1
